refactor memory mapping and memory buffering methods into separate file
