+++ src/lib/openjp2/CMakeLists.txt
@@ -37,6 +37,8 @@ set(OPENJPEG_SRCS
   ${CMAKE_CURRENT_SOURCE_DIR}/j2k.h
   ${CMAKE_CURRENT_SOURCE_DIR}/jp2.c
   ${CMAKE_CURRENT_SOURCE_DIR}/jp2.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/mem_stream.c
+  ${CMAKE_CURRENT_SOURCE_DIR}/mem_stream.h
   ${CMAKE_CURRENT_SOURCE_DIR}/mct.c
   ${CMAKE_CURRENT_SOURCE_DIR}/mct.h
   ${CMAKE_CURRENT_SOURCE_DIR}/mqc.c
